# Conflicts:
#	bundles/CoreBundle/src/Command/Bundle/InstallCommand.php
#	bundles/CoreBundle/src/Command/Bundle/ListCommand.php

+++ models/DataObject/Localizedfield/Dao.php
@@ -16,6 +16,7 @@
 namespace Pimcore\Model\DataObject\Localizedfield;
 
 use Doctrine\DBAL\Exception\TableNotFoundException;
+use Exception;
 use Pimcore\Db;
 use Pimcore\Db\Helper;
 use Pimcore\Logger;
@@ -77,7 +78,7 @@ class Dao extends Model\Dao\AbstractDao
 
     /**
      *
-     * @throws \Exception
+     * @throws Exception
      */
     public function save(array $params = []): void
     {
@@ -106,7 +107,7 @@ class Dao extends Model\Dao\AbstractDao
         }
 
         if (!isset($params['owner'])) {
-            throw new \Exception('need owner from container implementation');
+            throw new Exception('need owner from container implementation');
         }
 
         $this->model->_setOwner($params['owner']);
@@ -220,7 +221,7 @@ class Dao extends Model\Dao\AbstractDao
                 // if the table doesn't exist -> create it! deferred creation for object bricks ...
                 try {
                     $this->db->rollBack();
-                } catch (\Exception $er) {
+                } catch (Exception $er) {
                     // PDO adapter throws exceptions if rollback fails
                     Logger::info((string) $er);
                 }
@@ -261,7 +262,7 @@ class Dao extends Model\Dao\AbstractDao
                     // by the following DDL
                     try {
                         $this->db->rollBack();
-                    } catch (\Exception $er) {
+                    } catch (Exception $er) {
                         // PDO adapter throws exceptions if rollback fails
                         Logger::info((string) $er);
                     }
@@ -488,13 +489,13 @@ class Dao extends Model\Dao\AbstractDao
                     $fd->delete($object, $params);
                 }
             }
-        } catch (\Exception $e) {
+        } catch (Exception $e) {
             Logger::error((string) $e);
 
             if ($isUpdate && $e instanceof TableNotFoundException) {
                 try {
                     $this->db->rollBack();
-                } catch (\Exception $er) {
+                } catch (Exception $er) {
                     // PDO adapter throws exceptions if rollback fails
                     Logger::info((string) $er);
                 }
@@ -543,7 +544,7 @@ class Dao extends Model\Dao\AbstractDao
             $index = $context['index'] ?? $context['containerKey'] ?? null;
             $containerName = $context['fieldname'];
             if (!$context['containerType']) {
-                throw new \Exception('no container type set');
+                throw new Exception('no container type set');
             }
 
             $sql = Helper::quoteInto($this->db, 'src_id = ?', $objectId)." AND ownertype = 'localizedfield' AND "
@@ -552,13 +553,22 @@ class Dao extends Model\Dao\AbstractDao
                     '/'.$context['containerType'].'~'.$containerName.'/'.$index.'/%'
                 ).$dirtyLanguageCondition;
 
-            $this->db->executeStatement('DELETE FROM object_relations_'.$object->getClassId() . ' WHERE ' . $sql);
+            if ($deleteQuery || $context['containerType'] === 'fieldcollection') {
+                // Fieldcollection don't support delta updates, so we delete the relations and insert them later again
+                $this->db->executeStatement('DELETE FROM object_relations_'.$object->getClassId().' WHERE '.$sql);
+            }
 
             return true;
         }
 
-        $sql = 'ownertype = "localizedfield" AND ownername = "localizedfield" and src_id = '.$this->model->getObject()->getId().$dirtyLanguageCondition;
-        $this->db->executeStatement('DELETE FROM object_relations_'.$this->model->getObject()->getClassId() . ' WHERE ' . $sql);
+        if ($deleteQuery || $context['containerType'] === 'fieldcollection') {
+            // Fieldcollection don't support delta updates, so we delete the relations and insert them later again
+            $sql = 'ownertype = "localizedfield" AND ownername = "localizedfield" and src_id = '.$this->model->getObject(
+            )->getId().$dirtyLanguageCondition;
+            $this->db->executeStatement(
+                'DELETE FROM object_relations_'.$this->model->getObject()->getClassId().' WHERE '.$sql
+            );
+        }
 
         return false;
     }
@@ -616,7 +626,7 @@ class Dao extends Model\Dao\AbstractDao
         }
 
         if (!isset($params['owner'])) {
-            throw new \Exception('need owner from container implementation');
+            throw new Exception('need owner from container implementation');
         }
 
         $this->model->_setOwner($params['owner']);
@@ -765,7 +775,7 @@ QUERY;
 
                 // execute
                 $this->db->executeQuery($viewQuery);
-            } catch (\Exception $e) {
+            } catch (Exception $e) {
                 Logger::error((string) $e);
             }
         }
@@ -773,7 +783,7 @@ QUERY;
 
     /**
      *
-     * @throws \Exception
+     * @throws Exception
      */
     public function createUpdateTable(array $params = []): void
     {
